packages/hu-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "huap-core"
version = "0.1.0b1"
description = "HUAP Core - Build deterministic, traceable AI agents with replay and evaluation"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "HUAP Team"}
]
keywords = ["ai", "agents", "tracing", "llm", "evaluation", "replay"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/huap-ai/huap-core"
Documentation = "https://github.com/huap-ai/huap-core#readme"
Repository = "https://github.com/huap-ai/huap-core"

[project.scripts]
huap = "hu_core.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
encryption = [
    "cryptography>=41.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["hu_core*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
